updates for OTA fw update

+++ include/crc.h
@@ -1,5 +1,5 @@
 #include <stdint.h>
 
-uint8_t crc_crc8(const uint8_t *p, uint8_t len);
-uint32_t crc_crc32(const uint8_t *p, uint8_t len);
+uint8_t crc_crc8(const uint8_t *p, uint16_t len);
+uint32_t crc_crc32(const uint8_t *p, uint16_t len);
